Make image->{height, width} size_t

+++ include/cub3d/render/image.h
@@ -6,20 +6,22 @@
 /*   By: bgenia <bgenia@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/12 03:29:52 by bgenia            #+#    #+#             */
-/*   Updated: 2022/03/12 04:15:04 by bgenia           ###   ########.fr       */
+/*   Updated: 2022/03/12 04:58:46 by bgenia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef IMAGE_H
 # define IMAGE_H
 
+# include <stddef.h>
+
 # include <cub3d/render/color.h>
 
 typedef struct s_image
 {
 	void	*mlx;
-	int		width;
-	int		height;
+	size_t	width;
+	size_t	height;
 	void	*mlx_image;
 	char	*data;
 	int		bits_per_pixel;
@@ -28,11 +30,11 @@ typedef struct s_image
 }	t_image;
 
 t_image
-image_create(void *mlx, int width, int height);
+image_create(void *mlx, size_t width, size_t height);
 void
 image_destroy(t_image *image);
 
 t_color
-*image_get_pixel(t_image *image, int x, int y);
+*image_get_pixel(t_image *image, size_t x, size_t y);
 
 #endif
